+++ examples/load-more-infinite-scroll/pages/index.js
@@ -23,7 +23,7 @@ export default () => {
 
   return (
     <div>
-      <h1>Pagination</h1>
+      <h1>Infinite Loading</h1>
       {status === 'loading' ? (
         <p>Loading...</p>
       ) : status === 'error' ? (